move error to `error` module
