+++ src/eval/for_expr.rs
@@ -1,5 +1,4 @@
 use super::*;
-use crate::{template::ForDirective, ForExpr, Identifier, Object};
 
 pub(super) struct Collection<'a> {
     ctx: &'a Context<'a>,
