+++ src/eval/mod.rs
@@ -1,5 +1,6 @@
 //! HCL expression evaluation.
 
+mod error;
 mod expr;
 mod for_expr;
 mod func;
@@ -8,115 +9,15 @@ mod template;
 #[cfg(test)]
 mod tests;
 
-use self::func::{Func, FuncArgs};
-use crate::{
-    BinaryOperator, Error, Expression, Identifier, Map, ObjectKey, Result, UnaryOperator, Value,
-};
-use std::fmt;
+pub use self::error::{EvalError, EvalErrorKind, EvalResult};
+use self::for_expr::Collection;
+pub use self::func::{Func, FuncArgs};
+use crate::structure::*;
+use crate::template::*;
+use crate::{Error, Map, Number, Result, Value};
 
 mod private {
     pub trait Sealed {}
-}
-
-pub type EvalResult<T> = Result<T, EvalError>;
-
-#[derive(Debug)]
-pub struct EvalError {
-    inner: Box<EvalErrorKind>,
-}
-
-impl EvalError {
-    pub fn new(inner: EvalErrorKind) -> EvalError {
-        EvalError {
-            inner: Box::new(inner),
-        }
-    }
-
-    pub fn kind(&self) -> &EvalErrorKind {
-        &self.inner
-    }
-
-    fn unexpected<T>(value: T, expected: &'static str) -> EvalError
-    where
-        T: Into<Expression>,
-    {
-        EvalError::new(EvalErrorKind::Unexpected(value.into(), expected))
-    }
-}
-
-impl fmt::Display for EvalError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        fmt::Display::fmt(&self.inner, f)
-    }
-}
-
-impl From<&str> for EvalError {
-    fn from(msg: &str) -> Self {
-        EvalError::from(msg.to_string())
-    }
-}
-
-impl From<String> for EvalError {
-    fn from(msg: String) -> Self {
-        EvalError::new(EvalErrorKind::Message(msg))
-    }
-}
-
-impl From<Error> for EvalError {
-    fn from(err: Error) -> Self {
-        EvalError::from(err.to_string())
-    }
-}
-
-impl std::error::Error for EvalError {}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-#[non_exhaustive]
-pub enum EvalErrorKind {
-    RawExpression,
-    Message(String),
-    UndefinedVariable(Identifier),
-    UndefinedFunc(Identifier),
-    Unexpected(Expression, &'static str),
-    IndexOutOfBounds(usize),
-    InvalidUnaryOp(UnaryOperator, Expression),
-    InvalidBinaryOp(Expression, BinaryOperator, Expression),
-    NoSuchKey(ObjectKey),
-    KeyAlreadyExists(ObjectKey),
-}
-
-impl fmt::Display for EvalErrorKind {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            EvalErrorKind::RawExpression => f.write_str("raw expressions cannot be evaluated"),
-            EvalErrorKind::Message(msg) => f.write_str(msg),
-            EvalErrorKind::UndefinedVariable(ident) => {
-                write!(f, "undefined variable `{}`", ident)
-            }
-            EvalErrorKind::UndefinedFunc(ident) => {
-                write!(f, "undefined function `{}`", ident)
-            }
-            EvalErrorKind::Unexpected(expr, expected) => {
-                write!(f, "unexpected expression `{}`, expected {}", expr, expected)
-            }
-            EvalErrorKind::IndexOutOfBounds(index) => write!(f, "index out of bounds: {}", index),
-            EvalErrorKind::NoSuchKey(key) => write!(f, "no such key: `{}`", key),
-            EvalErrorKind::KeyAlreadyExists(key) => write!(f, "key `{}` already exists", key),
-            EvalErrorKind::InvalidUnaryOp(operator, expr) => write!(
-                f,
-                "unary operator `{}` is not applicable to `{}`",
-                operator.as_str(),
-                expr,
-            ),
-            EvalErrorKind::InvalidBinaryOp(lhs, operator, rhs) => write!(
-                f,
-                "binary operator `{}` is not applicable to `{}` and `{}`",
-                operator.as_str(),
-                lhs,
-                rhs
-            ),
-        }
-    }
 }
 
 /// A trait for evaluating HCL expressions.
